# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/documents/patient/{patientId}": {
      "get": {
        "operationId": "DocumentsController_list",
        "parameters": [
          {
            "name": "patientId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "documents"
        ]
      }
    },
    "/api/v1/documents/{id}": {
      "get": {
        "operationId": "DocumentsController_download",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "documents"
        ]
      }
    },
...
    "/api/v1/hr/staff/{userId}": {
      "get": {
        "operationId": "HrController_profileForUser",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "hr"
        ]
      }
    },
...
    "/api/v1/hr/leave/{id}": {
      "patch": {
        "operationId": "HrController_leaveStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "hr"
        ]
      }
    },
    "/api/v1/hr/staff/{id}/qualifications": {
      "patch": {
        "operationId": "HrController_setQualifications",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "hr"
        ]
      }
    },
...
  },
  "info": {
    "title": "FICMS API",
    "description": "Fertility & IVF Clinic Management System — versioned REST API",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
...
  }
}